+++ src/widget/mod.rs
@@ -32,5 +32,8 @@ pub mod image_widget;
 /// when appropriate.
 pub mod push_button_widget;
 
+/// Progress component: draws a progress meter widget.
+pub mod progress_widget;
+
 /// `Configurable` definition, used by `Widget` objects to store configuration settings.
 pub mod config;